Make the page look like the desired outcome
Made some changes to the html and css to make the page look like the
desired outcome using flex as per the instructions. I had it looking
like the desired outcome pic earlier but it did not wrap the text as
it should have, so I had to adjust where some of the flexboxes where
and what was in them. And I had to adjust the css some.